Update for my needs.
The original repo is an awesome idea, but the implementation didn't work
well for me. It seems the original author had tailored the script to his
particular use case which doesn't happen to mesh with my own. This
update seems to be more general and in any case works for me.